fix(dashboard): added missing turnstile
