+++ .changeset/late-rivers-wish.md
@@ -1,0 +1,5 @@
+---
+"@yukinu/dashboard": patch
+---
+
+added missing turnstile
